# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Margem.AI - Financial Decision Assistant
  description: |
    Margem.AI REST API, a financial management assistant aimed at
    Brazilian Individual Microentrepreneurs (MEIs). The API provides
    automated pricing calculations, profit margin, cash flow,
    and financial health monitoring, helping MEIs make informed
    business decisions.

    ## Modules
    - **Authentication**: registration, login, and JWT token renewal
    - **MEI Profile**: registration data and entrepreneur settings
    - **Fixed and Variable Costs**: registration and management of business costs
    - **Products and Services**: product/service catalog with base prices
    - **Sales**: sales registration with payment methods
    - **Pricing**: selling price calculator based on Markup
    - **Dashboard**: cash flow and financial summary
    - **Alerts**: MEI revenue cap monitoring
  version: 1.0.0
  license:
    name: MIT License
    identifier: MIT
servers:
  - url: https://api.margem.ai/v1
    description: Production
  - url: https://staging-api.margem.ai/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /categories:
    get:
      tags:
        - Categories
      summary: List categories
      description: Returns the MEI category catalog with optional filters and pagination.
      operationId: listCategories
      parameters:
        - name: type
          in: query
          description: Filter by type (product or service)
          schema:
            $ref: '#/components/schemas/ProductType'
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: search
          in: query
          description: Partial search by name
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCategories'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Categories
      summary: Create category
      description: Registers a dynamic category with inheritable operational parameters.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /categories/{id}:
    get:
      tags:
        - Categories
      summary: Get category by ID
      operationId: getCategory
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Categories
      summary: Update category
      operationId: updateCategory
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Categories
      summary: Deactivate category
      description: Soft-deletes the category. Blocked when there are active products or subcategories linked to it.
      operationId: removeCategory
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Category deactivated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /categories/{id}/status:
    patch:
      tags:
        - Categories
      summary: Update category status
      operationId: updateCategoryStatus
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryStatusRequest'
      responses:
        '200':
          description: Category status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /categories/{id}/revalidate-prices:
    post:
      tags:
        - Categories
      summary: Revalidate category prices in batch
      description: Recalculates the selling price of every active product in the category using the inherited parameters.
      operationId: revalidateCategoryPrices
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Batch revalidation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRevalidationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /settings/operational/fixed-costs-summary:
    get:
      tags:
        - Operational Settings
      summary: Get consolidated fixed costs summary
      description: Returns the cached total of active fixed costs, the revenue baseline and the resulting allocation rate (R_cf).
      operationId: getFixedCostsSummary
      responses:
        '200':
          description: Fixed costs summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixedCostsSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /settings/operational/rateio-config:
    patch:
      tags:
        - Operational Settings
      summary: Configure fixed cost allocation rate
      description: Configures the revenue baseline mode (target or historical) and whether R_cf is automatic or manually defined.
      operationId: updateRateioConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateioConfigRequest'
      responses:
        '200':
          description: Updated fixed costs summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixedCostsSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the authentication endpoint
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Unique resource identifier (UUID)
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "VALIDATION_ERROR"
            message: "Invalid data in request"
            details:
              - field: "email"
                message: "Invalid email format"
    Unauthorized:
      description: Not authenticated or invalid/expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired authentication token"
    Forbidden:
      description: Authenticated but not authorized to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "FORBIDDEN"
            message: "You do not have permission to access this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "NOT_FOUND"
            message: "Resource not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "INTERNAL_ERROR"
            message: "An unexpected error occurred"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "RATE_LIMIT_EXCEEDED"
            message: "Too many requests. Please try again later."
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error identifier code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Descriptive error message
          example: "Invalid data in request"
        details:
          type: array
          description: List of field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Invalid email format"
        timestamp:
          type: string
          format: date-time
...
    ProductType:
      type: string
      description: Item type (physical product or service)
      enum:
        - PRODUTO
        - SERVICO
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: "Bebidas"
        slug:
          type: string
          readOnly: true
          example: "bebidas"
        type:
          $ref: '#/components/schemas/ProductType'
        targetProfitMargin:
          type: number
          minimum: 0.01
          maximum: 99.99
          description: Default desired profit margin inherited by products without an explicit value
          example: 25.00
        taxRate:
          type: number
          minimum: 0
          maximum: 99.99
          description: Default tax rate inherited by products
          example: 6.00
        maxDiscountAllowed:
          type: number
          minimum: 0
          maximum: 100
          description: Maximum discount allowed for the category
          example: 10.00
        variableCostPercent:
          type: number
          minimum: 0
          maximum: 99.99
          description: Default variable sales costs / card fees percentage inherited by products
          example: 5.00
        active:
          type: boolean
          default: true
        parentId:
          type: string
          format: uuid
          nullable: true
        parentName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    CategoryRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
          example: "Bebidas"
        type:
          $ref: '#/components/schemas/ProductType'
        targetProfitMargin:
          type: number
          minimum: 0.01
          maximum: 99.99
          example: 25.00
        taxRate:
          type: number
          minimum: 0
          maximum: 99.99
          example: 6.00
        maxDiscountAllowed:
          type: number
          minimum: 0
          maximum: 100
          example: 10.00
        variableCostPercent:
          type: number
          minimum: 0
          maximum: 99.99
          example: 5.00
        parentId:
          type: string
          format: uuid
          nullable: true
    CategoryStatusRequest:
      type: object
      additionalProperties: false
      required:
        - active
      properties:
        active:
          type: boolean
    CategoryRevalidationResponse:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        updatedProducts:
          type: integer
          example: 3
    PaginatedCategories:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
    RevenueBaselineMode:
      type: string
      description: Source of the monthly revenue baseline used to compute R_cf
      enum:
        - TARGET_REVENUE
        - HISTORICAL_AVERAGE
    FixedCostsSummary:
      type: object
      properties:
        totalFixedCosts:
          type: number
          x-currency: BRL
          description: Cached sum of all active fixed costs
          example: 1000.00
        revenueBaseline:
          type: number
          x-currency: BRL
          description: Monthly revenue baseline considered
          example: 2000.00
        revenueBaselineMode:
          $ref: '#/components/schemas/RevenueBaselineMode'
        allocatedFixedCostPercent:
          type: number
          description: Resulting fixed cost allocation rate (R_cf)
          example: 50.00
        automaticRateio:
          type: boolean
        severeRisk:
          type: boolean
          description: True when fixed costs exceed 80% of the projected revenue
        riskMessage:
          type: string
        lastRecalculatedAt:
          type: string
          format: date-time
    RateioConfigRequest:
      type: object
      additionalProperties: false
      properties:
        revenueBaselineMode:
          $ref: '#/components/schemas/RevenueBaselineMode'
        monthlyRevenueTarget:
          type: number
          x-currency: BRL
          minimum: 0
          example: 5000.00
        historicalAverageRevenue:
          type: number
          x-currency: BRL
          minimum: 0
        automaticRateio:
          type: boolean
        manualAllocatedFixedCostPercent:
          type: number
          minimum: 0
          maximum: 99.99
...